Moved recording models into thier own namespace
