+++ src/HttpWebRequestWrapper.HttpClient/Extensions/HttpClientHandlerExtensions.cs
@@ -2,7 +2,6 @@
 using System.Net;
 using System.Net.Http;
 using System.Reflection;
-using System.Threading.Tasks;
 
 namespace HttpWebRequestWrapper.HttpClient.Extensions
 {
