+++ src/HttpWebRequestWrapper.Tests/InterceptorTests.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using HttpWebRequestWrapper.Recording;
 using HttpWebRequestWrapper.Tests.Properties;
 using Newtonsoft.Json;
 using Should;
