+++ src/HttpWebRequestWrapper/Extensions/RecordedRequestExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net;
+using HttpWebRequestWrapper.Recording;
 
 namespace HttpWebRequestWrapper.Extensions
 {
