+++ src/HttpWebRequestWrapper/HttpWebRequestWrapper.csproj
@@ -61,8 +61,11 @@
     <Compile Include="IInterceptorRequestBuilder.cs" />
     <Compile Include="InterceptedRequest.cs" />
     <Compile Include="IO\MemoryStreamExtensions.cs" />
-    <Compile Include="RecordedRequest.cs" />
-    <Compile Include="RecordingSession.cs" />
+    <Compile Include="Recording\RecordedHeaders.cs" />
+    <Compile Include="Recording\RecordedRequest.cs" />
+    <Compile Include="Recording\RecordedResponseException.cs" />
+    <Compile Include="Recording\RecordedStream.cs" />
+    <Compile Include="Recording\RecordingSession.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="IO\ShadowCopyStream.cs" />
     <Compile Include="RecordingSessionInterceptorRequestBuilder.cs" />
