+++ src/HttpWebRequestWrapper/HttpWebRequestWrapperInterceptor.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Net;
 using System.Reflection;
 using System.Threading;
+using HttpWebRequestWrapper.Recording;
 
 namespace HttpWebRequestWrapper
 {
